+++ src/lib/user/syscall.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <debug.h>
+#include <stdint.h>
 
 /* Process identifier. */
 typedef int pid_t;
@@ -54,6 +55,6 @@ int getuid (void);
 int geteuid (void);
 int getgid (void);
 int getegid (void);
-int chmod (const char *path, uint8_t, uint8_t, uint8_t);
+int chmod (const char *path, int, int, int);
 
 #endif /* lib/user/syscall.h */
